OS - Memoryの循環参照を解消
